add a variant 2 - use an integer hash on h2.
With the h2 to index function we have been using:

    idx= (h2 ^ xor_val) % num_buckets

we have the problem that when num_buckets is a power of 2 the
high bits of h2 and xor_val do not play any role in the final
result.

This means that for such hashes it is not unlikely we will have
an unresolvable collision and will have to generate a new seed
and restart from scratch. The entire existing "max_xor_val"
logic is there to allow us to short circuit out of such cases.

This patch adds a new variant 2 that uses a simple integer hash
on the result of h2 ^ xor_val, mixing the high and low bits
before computing the mod. This much reduces the chance that there
is an unresolvable collision, by allowing us to twiddle ALL the
bits in the h2.

We still have to guard against such unresolvable collisions by
trying a different seed to the hash function, but this change
much reduces the chance of them occuring.