support sorted keys on disk
This introduces variant 6 (no docs yet) which adds another
layer of indirection to the hash buckets so that the keys
are alphabetically ordered. This also changes the order that
will be output by the canonical form, using the same ordering
that perl would use for sort %keys. With this variant
keys() and each() will return the results in a sorted order.