Add ThreeLevelOnDisk, collection of improvements
ThreeLevelOnDisk (variant 7) is a version of MultiLevelOnDisk
that is optimised for 3 level HoHoH using composite keys. It
also supports compression of keys and/or values, and when compression
is not employed uses a trigram index during construction to detect
and exploit overlapping strings (which can be very common in some
data sets) to reduce memory footprint without compromising on run
time speed.

This includes various minor changes and tweaks for performace of
both MultiLevelOnDisk and ThreeLevelOnDisk, including the creation
of a container class Mount.pm which holds the mount object
independently from any ties that point at it. (This needs to be
"de-perled" and switched to use IV_TO_PTR or equivalent.)